requirements.txt
alembic==1.9.4
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
click==8.1.3
coverage==7.2.1
exceptiongroup==1.1.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.1.0
pluggy==1.0.0
pydantic==1.10.5
PyJWT==2.6.0
pytest==7.2.2
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
